refactor(stories): Combine stories for UI subcomponents.
