+++ packages/trip-form/src/__mocks__/mode-icon-wrap.js
